backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
Pillow==10.1.0
pytesseract==0.3.10
google-cloud-vision==3.4.4
gunicorn==21.2.0
python-dotenv==1.0.0